requirements.txt

pickle
pandas
scikit-learn
Pillow
streamlit
